Fix translation
