+++ flash/text/seq2seq/translation/model.py
@@ -28,7 +28,7 @@ class TranslationTask(Seq2SeqTask):
         loss_fn: Loss function for training.
         optimizer: Optimizer to use for training, defaults to `torch.optim.Adam`.
         metrics: Metrics to compute for training and evaluation.
-        learning_rate: Learning rate to use for training, defaults to `3e-4`
+        learning_rate: Learning rate to use for training, defaults to `1e-5`
         val_target_max_length: Maximum length of targets in validation. Defaults to `128`
         num_beams: Number of beams to use in validation when generating predictions. Defaults to `4`
         n_gram: Maximum n_grams to use in metric calculation. Defaults to `4`
@@ -41,11 +41,11 @@ class TranslationTask(Seq2SeqTask):
         loss_fn: Optional[Union[Callable, Mapping, Sequence]] = None,
         optimizer: Type[torch.optim.Optimizer] = torch.optim.Adam,
         metrics: Union[pl.metrics.Metric, Mapping, Sequence, None] = None,
-        learning_rate: float = 3e-4,
+        learning_rate: float = 1e-5,
         val_target_max_length: Optional[int] = 128,
         num_beams: Optional[int] = 4,
         n_gram: bool = 4,
-        smooth: bool = False,
+        smooth: bool = True,
     ):
         self.save_hyperparameters()
         super().__init__(
@@ -70,11 +70,11 @@ class TranslationTask(Seq2SeqTask):
         tgt_lns = self.tokenize_labels(batch["labels"])
         # wrap targets in list as score expects a list of potential references
         tgt_lns = [[reference] for reference in tgt_lns]
-        result = self.bleu(generated_tokens, tgt_lns)
+        result = self.bleu(self._postprocess.uncollate(generated_tokens), tgt_lns)
         self.log(f"{prefix}_bleu_score", result, on_step=False, on_epoch=True, prog_bar=True)
 
     def _ci_benchmark_fn(self, history: List[Dict[str, Any]]):
         """
         This function is used only for debugging usage with CI
         """
-        # assert history[-1]["val_bleu_score"]
+        assert history[-1]["val_bleu_score"] > 0.6
